refactor: rename files and commmands
-script.py is rename to cli.py, and  in the commands suffix has been removed
