+++ setup.py
@@ -84,7 +84,7 @@ setup(
     },
     entry_points={
         "console_scripts": [
-            "shopyo=shopyo.api.scripts:cli"
+            "shopyo=shopyo.api.cli:cli"
         ]
     },
 )
