+++ shopyo/manage.py
@@ -5,7 +5,7 @@ Example all ``shopyo <cmd> [OPTIONS] [ARGS]...`` commands can also be run as
 ``python manage.py <cmd> [OPTIONS] [ARGS]...``
 """
 
-from shopyo.api.scripts import cli
+from shopyo.api.cli import cli
 
 if __name__ == '__main__':
     cli()
